+++ samples/basic/random_walk_trace_layer.rb
@@ -34,17 +34,33 @@ end
 class Screen < Nyle::Screen
 
   def initialize
-    super(320, 240, {bgcolor: :BLACK, trace: true})
-    @walker = Walker.new(@width, @height)
+    super(320, 240, {bgcolor: :BLACK, trace: false})
+    @walker1 = Walker.new(@width, @height, :YELLOW)
+    @walker2 = Walker.new(@width, @height, :GREEN)
+    @walker3 = Walker.new(@width, @height, :BLUE)
+    @walker4 = Walker.new(@width, @height, :RED)
   end
 
   def draw
-    @walker.draw
-    Nyle.draw_rect(0, 0, @width, @height, {fill: true, color: :BLACK, a: 0.02})
+    Nyle.layer(:L1) do
+      @walker1.draw
+      Nyle.draw_rect(0, 0, @width, @height, {fill: true, color: :BLACK, a: 0.02})
+    end
+    Nyle.layer(:L2) do
+      @walker2.draw
+    end
+    Nyle.layer(:L3) do
+      Nyle.clear
+      @walker3.draw
+    end
+    @walker4.draw
   end
 
   def update
-    @walker.move
+    @walker1.move
+    @walker2.move
+    @walker3.move
+    @walker4.move
     Nyle.quit if Nyle.key_press?(KEY_Escape)
   end